Change factorial from int to long long

+++ hmLib/math.hpp
@@ -6,7 +6,7 @@ namespace hmLib{
 		constexpr double pi = 3.14159265358979323846;
 		constexpr double napier = 2.71828182845904523536;
 		
-		inline unsigned int factorial(unsigned int val){
+		inline unsigned long long factorial(unsigned long long val){
 			return val>1 ? val*factorial(val - 1) : 1;
 		}
 	}
